pyproject.toml
[project]
name = "sgcif-vae"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=6.30.1",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "prettytable>=3.16.0",
    "pyyaml>=6.0.2",
    "torch>=2.8.0",
    "torchvision>=0.23.0",
    "tqdm>=4.67.1",
]
